+++ tests/test_satellite.py
@@ -29,3 +29,19 @@ class TestSmoke(unittest.TestCase):
 
         for constellation in eval(value):
             self.assertIsInstance(constellation, frozenset)
+
+    def test_satellite(self):
+        """Verify contents of the output"""
+
+        file_path = os.path.join(example_dir, 'satellite.py')
+        output = subprocess.check_output([sys.executable, file_path])
+        output = str(output).upper()
+        if os.getenv('DEBUG_OUTPUT'):
+            print("Example output \n"+ output)
+
+        with self.subTest(msg="Verify if output contains '[frozenset({' \n"):
+            self.assertIn("[frozenset({".upper(), output)
+        with self.subTest(msg="Verify if error string contains in output \n"):
+            self.assertNotIn("ERROR", output)
+        with self.subTest(msg="Verify if warning string contains in output \n"):
+            self.assertNotIn("WARNING", output)
